# chapter-6/final-lab/jobposting-back/react-openapi-generator/swagger.yaml
openapi: 3.0.1
info:
  title: Jobposting API
  description: Jobposting API documentation
  termsOfService: ''
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/applications/{applicationId}:
    put:
      tags:
        - application-resource
      operationId: updateApplication
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Application'
    patch:
      tags:
        - application-resource
      operationId: partialUpdateApplication
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/Application'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Application'
...
  /api/applications:
    get:
      tags:
        - application-resource
      operationId: getAllApplications
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
    post:
      tags:
        - application-resource
      operationId: createApplication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Application'
...
  /api/applications/{id}:
    get:
      tags:
        - application-resource
      operationId: getApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Application'
    delete:
      tags:
        - application-resource
      operationId: deleteApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
...
components:
  schemas:
    Applicant:
      type: object
      properties:
        applicantId:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        phone:
          type: string
    Application:
      type: object
      properties:
        applicationId:
          type: integer
          format: int64
        applicationDate:
          type: string
          format: date-time
        resume:
          $ref: '#/components/schemas/Resume'
        jobPosting:
          $ref: '#/components/schemas/JobPosting'
        applicant:
          $ref: '#/components/schemas/Applicant'
    JobPosting:
      type: object
      properties:
        jobId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        location:
          type: string
        postedDate:
          type: string
          format: date-time
    Resume:
      type: object
      properties:
        resumeId:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        submittedDate:
          type: string
          format: date-time
        applications:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Application'
...